requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
dotenv
gunicorn
idna
packaging
psycopg2-binary
PyJWT
python-dotenv
requests
sqlparse
stripe
typing_extensions
urllib3
channels
channels_redis
uvicorn
daphne
django-redis
celery[redis]
redis